Move js to .js
